Add a warning when the ctlscript is executed as root
It can happen that a user wants to install/update the configuration
files with scope "all" (which is the default) and then executes the
ctlscript script using sudo. In that case the user-specific files
that are going to be installed/updated are those that belongs to
the user root. Most probably that is not desired. This commit adds
a warning to alert the user if the script is executed as root.